+++ crates/amaru/src/lib.rs
@@ -27,6 +27,7 @@ pub mod exit;
 pub mod metrics;
 pub mod observability;
 pub mod panic;
+pub mod peer_snapshot;
 pub mod stages;
 pub mod submit_api;
 pub mod version;
@@ -214,6 +215,9 @@ pub mod env_vars {
 
     /// --peer-address
     pub const PEER_ADDRESS: &str = "AMARU_PEER_ADDRESS";
+
+    /// --peer-snapshot
+    pub const PEER_SNAPSHOT: &str = "AMARU_PEER_SNAPSHOT";
 
     /// --peer-removal-cooldown-secs
     pub const PEER_REMOVAL_COOLDOWN_SECS: &str = "AMARU_PEER_REMOVAL_COOLDOWN_SECS";